refacto archicteture and create independent component for joinChannel
